requirements.txt
# =============================================================
# TenAsia Intelligence Hub — Python Dependencies
# Python 3.10+
# =============================================================

# ── Web Framework ─────────────────────────────────────────────
streamlit>=1.28.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── 환경 설정 ─────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── AI ────────────────────────────────────────────────────────
google-generativeai>=0.8.3

# ── Web / Scraping ────────────────────────────────────────────
beautifulsoup4>=4.12.0
lxml>=5.0.0               # bs4 파서 (html.parser 대비 빠름)
requests>=2.31.0
urllib3>=2.0.0

# ── Video ─────────────────────────────────────────────────────
yt-dlp>=2023.12.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# ── Database (ORM + 마이그레이션) ────────────────────────────
SQLAlchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9    # PostgreSQL 드라이버 (바이너리 포함)

# ── 로깅 ──────────────────────────────────────────────────────
structlog>=24.0.0

# ── AWS (Secrets Manager, S3, SSM) ───────────────────────────
boto3>=1.34.0
